bug fix on http delete all

+++ garganttua-api-spring/garganttua-api-spring-interfaces/garganttua-api-spring-interface-rest/src/main/java/com/garganttua/api/interfaces/spring/rest/GGAPIServiceMethodToHttpMethodBinder.java
@@ -31,7 +31,9 @@ public class GGAPIServiceMethodToHttpMethodBinder {
 				return GGAPIEntityOperation.read_one;
 			return GGAPIEntityOperation.read_all;
 		case "DELETE":
-			return GGAPIEntityOperation.delete_one;
+			if( endpoint.endsWith("*") )
+				return GGAPIEntityOperation.delete_one;
+			return GGAPIEntityOperation.delete_all;
 		case "PATCH":
 			return GGAPIEntityOperation.update_one;
 		case "PUT":
